The runtime manager was 2,850 lines, and three more guards read one file
The host half of the local runtime -- starting a VM on macOS, importing a
distribution on Windows, moving the data into its holder, requests to the
guest agent, diagnostics when none of it works -- was one `lib.rs`.

Ten modules now, split by what the host is doing to the guest rather than
by platform alone: `lifecycle`, `request`, `clock`, `guest_image`,
`diagnostics`, `private_files`, `wsl_command`, `windows`, `windows_data`
and `macos`. The inherent impl on `ManagedRuntime` is split across them.
`lib.rs` keeps the config, the status, the struct and `new`.

The 946-line test module became ten files grouped the same way.

Three guards read `include_str!("lib.rs")`, and all three broke. They are
the WIN-1 guards: that an upgrade will not delete a runtime still holding
the data, that the gate runs before the `--unregister` it guards, and that
the migration quiesces the distribution before copying it. So
`manager_source()` reads every module from disk and asserts it found more
than eight files, and a new guard fails any test that goes looking for the
manager inside `lib.rs`.

Two re-exports are deliberately narrow. `wsl_command` and `windows_data`
are `#[cfg]`-gated all the way down, so a plain `pub(crate) use` of either
is dead on the host that is not compiling it -- which `-D warnings` calls
an error, on the platform CI checks and this machine does not.
`windows_data` in particular is reached by bare name only from the guards,
so its re-export says `#[cfg(test)]` and nothing more.
